权限管理相关功能开发

+++ zblog-base/src/main/java/com/github/stazxr/zblog/base/component/security/handler/CustomLogoutHandler.java
@@ -28,14 +28,16 @@ public class CustomLogoutHandler implements LogoutHandler {
 
     @Override
     public void logout(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
-        User user = (User) authentication.getPrincipal();
-        String username = user.getUsername();
-        log.info("用户 {} 正在注销...", username);
+        if (authentication != null && authentication.getPrincipal() != null) {
+            User user = (User) authentication.getPrincipal();
+            String username = user.getUsername();
+            log.info("用户 {} 正在注销...", username);
 
-        // 注销token
-        jwtTokenStorage.expire(user.getId());
+            // 注销token
+            jwtTokenStorage.expire(user.getId());
 
-        // 清除RememberMe
-        zblogService.removeRememberMe(username);
+            // 清除RememberMe
+            zblogService.removeRememberMe(username);
+        }
     }
 }
